Finished separating file list
